CockpitDisplay & UI
updated XR Origin prefab and added vfx, audio, another sprite and adjusted positions
