+++ _OctaviCopter/Assets/Audio/TutorialVoice.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 57258004b83f84f038213b8c510f9ee5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
