+++ _OctaviCopter/Assets/Audio/TutorialVoice/Test.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 71c53848f498b45b9be3b5e41bc73205
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
