+++ _OctaviCopter/Assets/Audio/TutorialVoice/Test/_MarcoTest Kopie.mp3.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 1be29bba1a2c040fba4d27593b511c56
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
