+++ _OctaviCopter/Assets/Prefabs/BoltVisuals.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 614beb4aae13a4446bf811a6fe4e5441
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
