+++ _OctaviCopter/Assets/Prefabs/BoltVisuals/AudioInput_Base.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 27464438be0ed43a7b25f8db3c794a67
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
