+++ _OctaviCopter/Assets/Prefabs/BoltVisuals/MicroInputBase_Base.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 73b54d1e64a7140208412140e865e683
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
