+++ _OctaviCopter/Assets/Prefabs/MiniOctaviCopter/MiniOctavi_AdeepBlue.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 83624fc05ce6c46aeb07d2c8c7abe3aa
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
