+++ _OctaviCopter/Assets/Prefabs/MiniOctaviCopter/MiniOctavi_Dorange.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f97264dd7f15c46d3b04810387a20276
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
