+++ _OctaviCopter/Assets/Prefabs/MiniOctaviCopter/MiniOctavi_Eyellow.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 245943ac6655141e894cc268b0c26d79
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
