+++ _OctaviCopter/Assets/Prefabs/MiniOctaviCopter/MiniOctavi_Fgreen.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9a8de746785e647509adc7d49356be8a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
