+++ _OctaviCopter/Assets/Prefabs/MiniOctaviCopter/MiniOctavi_Gblue.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ff81a069dcaf54cb09fbd3d2a1f0589b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
