+++ _OctaviCopter/Assets/Scripts/BoltAudio/AudioData_AmplitudeBand.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 264496decef2c4f9dbac9a54e479129e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
