+++ _OctaviCopter/Assets/Scripts/BoltAudio/MicrophoneInput.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0a8ddba0497ac402085c1231323f162e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
