+++ _OctaviCopter/Assets/Scripts/Macros/AudioInputHandler_Tutorial 1.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ec609c2635d2d4bf180644b44e26e415
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
