--- _OctaviCopter/Assets/Scripts/Macros/AudioInputHandler_Tutorial.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2869d2a6a6337be43a871423d670c621
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
