+++ _OctaviCopter/Assets/Scripts/Macros/Audio_TutorialVoice.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2869d2a6a6337be43a871423d670c621
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
