+++ _OctaviCopter/Assets/Shader/VFXGraphs/BoltAudio.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6265a9e542bca40cea1dfe48f7c0529c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
