+++ _OctaviCopter/Assets/Shader/VFXGraphs/BoltAudio/AudioInput_Base 1.vfx.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 19f310abbc72e44a1ab9c4f8a5ed9cd7
+VisualEffectImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
