+++ _OctaviCopter/Assets/Shader/VFXGraphs/BoltAudio/TutorialVoice.vfx.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 2d7a47995391e4e3992b72c059251914
+VisualEffectImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
